cli/fine_tuned_model/src/requirements.txt
pandas==2.2.2
PyYAML==6.0.1
PyYAML==6.0.1
transformers==4.43.3
